requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-dotenv
email-validator
passlib[bcrypt]
pyjwt
python-multipart
alembic
psycopg2-binary